requirements/ci.txt
twine==1.9.1
flake8==3.5.0
isort==4.3.4
pytest==3.4.0
pytest-aiohttp==0.3.0
pytest-cov==2.5.1
sphinx==1.7.0
sphinxcontrib-asyncio==0.2.0
sphinxcontrib-spelling==4.0.1
-r prod.txt
